Minor edits for styling
